backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
pydantic
numpy
PyPDF2
python-docx
